Add a new column `Issue states` that can be sorted on
That way it is much easier to find FIXMEs for issues that are fixed,
because one can sort on issues that are "Closed".

I'm sure there is a way to present this in a better way, but this is
already a significant improvement, so is a reasonable starting point for
further tweaks.